refactor: studyInfoData fetching 관련 내용 훅으로 분리
